+++ r/R/dataset-write.R
@@ -21,12 +21,17 @@
 #' binary storage formats, and by specifying relevant partitioning, you can
 #' make it much faster to read and query.
 #'
-#' @param dataset [Dataset] or `arrow_dplyr_query`. If a `arrow_dplyr_query`,
-#' note that `select()` or `filter()` queries are not currently supported.
+#' @param dataset [Dataset], [RecordBatch], [Table], `arrow_dplyr_query`, or
+#' `data.frame`. If an `arrow_dplyr_query` or `grouped_df`,
+#' `schema` and `partitioning` will be taken from the result of any `select()`
+#' and `group_by()` operations done on the dataset. Note that `filter()` queries
+#' are not currently supported, and `select`-ed columns may not be renamed.
 #' @param path string path to a directory to write to (directory will be
 #' created if it does not exist)
 #' @param format file format to write the dataset to. Currently only "feather"
 #' (aka "ipc") is supported.
+#' @param schema [Schema] containing a subset of columns, possibly reordered,
+#' in `dataset`. Default is `dataset$schema`, i.e. all columns.
 #' @param partitioning `Partitioning` or a character vector of columns to
 #' use as partition keys (to be written as path segments). Default is to
 #' use the current `group_by()` columns.
@@ -38,25 +43,34 @@
 write_dataset <- function(dataset,
                           path,
                           format = dataset$format$type,
+                          schema = dataset$schema,
                           partitioning = dplyr::group_vars(dataset),
                           hive_style = TRUE,
                           ...) {
   if (inherits(dataset, "arrow_dplyr_query")) {
     force(partitioning) # get the group_vars before we drop the object
-    # TODO: Write a filtered/projected dataset
+    # Check for a filter
     if (!isTRUE(dataset$filtered_rows)) {
+      # TODO:
       stop("Writing a filtered dataset is not yet supported", call. = FALSE)
     }
+    # Check for a select
     if (!identical(dataset$selected_columns, set_names(names(dataset$.data)))) {
-      # TODO: actually, we can do this?
-      stop("TODO", call. = FALSE)
+      # We can select a subset of columns but we can't rename them
+      if (!setequal(dataset$selected_columns, names(dataset$selected_columns))) {
+        stop("Renaming columns when writing a dataset is not yet supported", call. = FALSE)
+      }
+      dataset <- ensure_group_vars(dataset)
+      schema <- dataset$.data$schema[dataset$selected_columns]
     }
     dataset <- dataset$.data
   }
+  if (inherits(dataset, c("data.frame", "RecordBatch", "Table"))) {
+    force(partitioning) # get the group_vars before we replace the object
+    dataset <- InMemoryDataset$create(dataset)
+  }
   if (!inherits(dataset, "Dataset")) {
-    stop("'dataset' must be a Dataset", call. = FALSE)
-    # TODO: This does not exist yet (in the R bindings at least)
-    # dataset <- InMemoryDataset$create(dataset)
+    stop("'dataset' must be a Dataset, not ", class(dataset)[1], call. = FALSE)
   }
 
   if (!inherits(format, "FileFormat")) {
@@ -78,5 +92,5 @@ write_dataset <- function(dataset,
       partitioning <- DirectoryPartitioning$create(partition_schema)
     }
   }
-  dataset$write(path, format = format, partitioning = partitioning, ...)
+  dataset$write(path, format = format, partitioning = partitioning, schema = schema, ...)
 }